+++ Packages/com.unity.inputsystem/InputSystem/Controls/AxisControl.cs
@@ -167,6 +167,8 @@ namespace UnityEngine.InputSystem.Controls
         /// <value>Multiplier for input values.</value>
         public float scaleFactor;
 
+        public override float magnitude => EvaluateMagnitude(value);
+
         /// <summary>
         /// Apply modifications to the given value according to the parameters configured
         /// on the control (<see cref="clamp"/>, <see cref="normalize"/>, etc).
@@ -257,25 +259,29 @@ namespace UnityEngine.InputSystem.Controls
         /// <inheritdoc />
         public override unsafe float EvaluateMagnitude(void* statePtr)
         {
-            var value = ReadValueFromState(statePtr);
+            return EvaluateMagnitude(ReadValueFromState(statePtr));
+        }
+
+        private float EvaluateMagnitude(float value)
+        {
             if (m_MinValue.isEmpty || m_MaxValue.isEmpty)
                 return Mathf.Abs(value);
 
             var min = m_MinValue.ToSingle();
             var max = m_MaxValue.ToSingle();
 
-            value = Mathf.Clamp(value, min, max);
+            var clampedValue = Mathf.Clamp(value, min, max);
 
             // If part of our range is in negative space, evaluate magnitude as two
             // separate subspaces.
             if (min < 0)
             {
-                if (value < 0)
-                    return NormalizeProcessor.Normalize(Mathf.Abs(value), 0, Mathf.Abs(min), 0);
-                return NormalizeProcessor.Normalize(value, 0, max, 0);
+                if (clampedValue < 0)
+                    return NormalizeProcessor.Normalize(Mathf.Abs(clampedValue), 0, Mathf.Abs(min), 0);
+                return NormalizeProcessor.Normalize(clampedValue, 0, max, 0);
             }
 
-            return NormalizeProcessor.Normalize(value, min, max, 0);
+            return NormalizeProcessor.Normalize(clampedValue, min, max, 0);
         }
     }
 }