+++ Packages/com.unity.inputsystem/InputSystem/Plugins/XR/XRSupport.cs
@@ -4,6 +4,7 @@ using System;
 using System.Collections.Generic;
 using UnityEngine.InputSystem.Layouts;
 using UnityEngine.InputSystem.Controls;
+using UnityEngine.InputSystem.LowLevel;
 using UnityEngine.XR;
 
 namespace UnityEngine.InputSystem.XR
@@ -142,15 +143,17 @@ namespace UnityEngine.InputSystem.XR
         /// <summary>
         /// The index with the device's controls array where the parent bone resides.
         /// </summary>
-        public uint parentBoneIndex { get; set; }
+        public uint parentBoneIndex;
+
         /// <summary>
         /// The tracked position of the bone.
         /// </summary>
-        public Vector3 position { get; set; }
+        public Vector3 position;
+
         /// <summary>
         /// The tracked rotation of the bone.
         /// </summary>
-        public Quaternion rotation { get; set; }
+        public Quaternion rotation;
     }
 
     /// <summary>
@@ -161,31 +164,31 @@ namespace UnityEngine.InputSystem.XR
         /// <summary>
         /// The tracked position of the left eye.
         /// </summary>
-        public Vector3 leftEyePosition { get; set; }
+        public Vector3 leftEyePosition;
         /// <summary>
         /// The tracked rotation of the left eye.
         /// </summary>
-        public Quaternion leftEyeRotation { get; set; }
+        public Quaternion leftEyeRotation;
         /// <summary>
         /// The tracked position of the right eye.
         /// </summary>
-        public Vector3 rightEyePosition { get; set; }
+        public Vector3 rightEyePosition;
         /// <summary>
         /// The tracked rotation of the right eye.
         /// </summary>
-        public Quaternion rightEyeRotation { get; set; }
+        public Quaternion rightEyeRotation;
         /// <summary>
         /// The point in 3D space that the pair of eyes is looking.
         /// </summary>
-        public Vector3 fixationPoint { get; set; }
+        public Vector3 fixationPoint;
         /// <summary>
         /// The amount [0-1] the left eye is open or closed.  1.0 is fully open.
         /// </summary>
-        public float leftEyeOpenAmount { get; set; }
+        public float leftEyeOpenAmount;
         /// <summary>
         /// The amount [0-1] the right eye is open or closed.  1.0 is fully open.
         /// </summary>
-        public float rightEyeOpenAmount { get; set; }
+        public float rightEyeOpenAmount;
     }
 
     public class BoneControl : InputControl<Bone>
@@ -196,6 +199,31 @@ namespace UnityEngine.InputSystem.XR
         public Vector3Control position { get; private set; }
         [InputControl(offset = 16, displayName = "Rotation")]
         public QuaternionControl rotation { get; private set; }
+
+        public override ref readonly Bone value
+        {
+            get
+            {
+#if UNITY_EDITOR
+                if (!useCachedValue)
+                    return ref ReadStateInEditor();
+#endif
+
+                if (!m_CachedValueIsStale) return ref m_CachedValue;
+
+                m_CachedValue = new Bone
+                {
+                    parentBoneIndex = (uint)parentBoneIndex.unprocessedValue,
+                    position = position.unprocessedValue,
+                    rotation = rotation.unprocessedValue
+                };
+
+                ProcessValue(ref m_CachedValue);
+                m_CachedValueIsStale = false;
+
+                return ref m_CachedValue;
+            }
+        }
 
         protected override void FinishSetup()
         {
@@ -277,6 +305,33 @@ namespace UnityEngine.InputSystem.XR
             fixationPoint.WriteValueIntoState(value.fixationPoint, statePtr);
             leftEyeOpenAmount.WriteValueIntoState(value.leftEyeOpenAmount, statePtr);
             rightEyeOpenAmount.WriteValueIntoState(value.rightEyeOpenAmount, statePtr);
+        }
+
+        public override ref readonly Eyes value
+        {
+            get
+            {
+#if UNITY_EDITOR
+                if (!useCachedValue)
+                    return ref ReadStateInEditor();
+#endif
+                if (!m_CachedValueIsStale) return ref m_CachedValue;
+
+                m_CachedValue = new Eyes
+                {
+                    leftEyePosition = leftEyePosition.unprocessedValue,
+                    leftEyeRotation = leftEyeRotation.unprocessedValue,
+                    rightEyePosition = rightEyePosition.unprocessedValue,
+                    rightEyeRotation = rightEyeRotation.unprocessedValue,
+                    fixationPoint = fixationPoint.unprocessedValue,
+                    leftEyeOpenAmount = leftEyeOpenAmount.unprocessedValue,
+                    rightEyeOpenAmount = rightEyeOpenAmount.unprocessedValue
+                };
+                ProcessValue(ref m_CachedValue);
+                m_CachedValueIsStale = false;
+
+                return ref m_CachedValue;
+            }
         }
     }
 #pragma warning restore 0649